Conflicts:
	src/Analyzer.h

+++ src/Particle.h
@@ -252,7 +252,7 @@ public:
   pair<vector<int>*,vector<int>* > maxIso = make_pair(nullptr,nullptr);
   vector<double>  *leadChargedCandPt = 0;
   vector<double>  *leadChargedCandPtError = 0;
-  vector<int>  *leadChargedCandValidHits = 0;
+  vector<double>  *leadChargedCandValidHits = 0;
 };
